refactor(Reservation): 예약 관련 로직 리팩토링(레이어드 아키텍처 반영)
