+++ src/main/java/roomescape/model/reservation/Reservation.java
@@ -2,7 +2,6 @@ package roomescape.model.reservation;
 
 import roomescape.model.time.Time;
 
-// 예약 관련 데이터베이스 엔티티
 public class Reservation {
 
     private Long id;
